add react router, add default route to v1
