+++ src/router.js
@@ -1,0 +1,16 @@
+
+import { createBrowserRouter, Navigate } from "react-router-dom";
+import HomeView from "./view/HomeView";
+
+const router = createBrowserRouter([
+  {
+    path: "/",
+    element: <Navigate to="/v1" />,
+  },
+  {
+    path: "/v1",
+    element: <HomeView />,
+  },
+]);
+
+export default router
